[REFACTORING] + [HF] `Added:` some jsDoc typing [HF] - `Changed:` QrScanner text and button centered
